Extracts `ManagedCache` and core data helpers to new file
